requirements.txt
Flask~=2.2.2
python-dotenv~=0.20.0
aiohttp~=3.8.1
beautifulsoup4~=4.11.1
Pillow~=9.2.0